Refactor: split monolithic datalayer_sqlite.py into focused submodules
- Created vultron/adapters/driven/datalayer_sqlite/ subpackage with semantic separation
- schema.py (106 lines): SQLModel table definitions and helpers
- engine.py (98 lines): Engine factory with WAL pragmas and JSON serialization
- crud.py (354 lines): Create, read, get, save, update, delete operations
- queries.py (318 lines): Query and search delegates (count, by_type, find_*)
- queues.py (242 lines): Inbox/outbox queue operations
- datalayer.py (566 lines): Main SqliteDataLayer class delegating to submodules
- __init__.py (154 lines): Public API re-exports and factory functions
- Replaced monolithic file with compatibility shim for backward compatibility
- All submodules follow project organization guidelines (CS-18-002)
- No behavior changes; all tests pass (3416 passed)
